Performance refactor: write 8 bytes rather than 4 at a time in DiehardInputGenerator. Reduce Diehard to 1/10 of default psamples.

+++ core/src/java/main/org/uncommons/maths/random/DiehardInputGenerator.java
@@ -72,7 +72,7 @@ public final class DiehardInputGenerator
         dataOutput = new DataOutputStream(new BufferedOutputStream(new FileOutputStream(outputFile)));
         try {
             while (true) {
-                dataOutput.writeInt(rng.nextInt());
+                dataOutput.writeLong(rng.nextLong());
             }
         } catch (IOException expected) {
             // Broken pipe means Dieharder is finished
